UI_01/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
bcrypt>=4.0.0
pyjwt>=2.6.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
livekit-api==0.7.1
livekit==0.11.0
groq
redis[asyncio]==5.0.4
aiokafka>=0.10.0
pytz
httpx>=0.27.0
# Optional — Whisper STT (IVR transcription). Install separately if needed:
# openai-whisper